+++ database/migrations/2017_09_25_062009_create_invoice_posts_table.php
@@ -15,9 +15,9 @@ class CreateInvoicePostsTable extends Migration
     {
         Schema::create('invoice_posts', function (Blueprint $table) {
             $table->id('id');
-            $table->integer('invoice_id')->unsigned()->nullable(); 
+            $table->bigInteger('invoice_id')->unsigned()->nullable(); 
             $table->string('video')->nullable();
-            $table->integer('postable_id')->nullable();
+            $table->bigInteger('postable_id')->nullable();
             $table->string('postable_type')->nullable();
             $table->date('date');
             $table->boolean('published')->default(false);
