+++ database/migrations/2017_10_18_080943_create_wedding_rings_table.php
@@ -29,7 +29,7 @@ class CreateWeddingRingsTable extends Migration
             $table->boolean('customized')->default(false);
             $table->string('gender')->default('m');
             $table->integer('page_id')->unsigned()->nullable();
-            $table->integer('wedding_ring_pair_id')->unsigned()->nullable();
+            $table->bigInteger('wedding_ring_pair_id')->unsigned()->nullable();
             $table->timestamps();
         });
