+++ database/migrations/2017_10_18_081127_create_pages_table.php
@@ -16,7 +16,7 @@ class CreatePagesTable extends Migration
         Schema::create('pages', function (Blueprint $table) {
             $table->id('id');
             $table->string('url');
-            $table->integer('paginable_id')->nullable(); 
+            $table->bigInteger('paginable_id')->nullable(); 
             $table->string('paginable_type')->nullable();  
             $table->timestamps();
         });
